requirements.txt
pandas
geopandas
dataset
python-dotenv
psycopg2
python-slugify
